teletext.c — Bug 1 Changed height from 32 to 24 and updated the comment block. This reduces output to 12 rows per field, mapping to VBI lines 10–21 exactly. The old value of 32 silently wrote into active video lines 22–25. demo.c — Bugs 2 and 3 Bug 2 — Two changes, both flowing from the same §3.3 requirement:
nl_byte() now calls parity(b) before storing into nl_page[]. Since every NAPLPS byte in demo_graphics() goes through this function, the entire graphics path is fixed at one point.
demo_ascii() had a separate memcpy path that bypassed nl_byte(). Replaced with a parity()-applying loop. Also removed the now-unused <string.h> memcpy call and added #include "hamming.h" at the top.

Bug 3 — Added a size guard at the top of push_page() that clamps nl_len to NABTS_FSS_MAX_NAPLPS (1896) with a stderr warning. The cap is defined by constants in nabts.h so callers can check against it themselves.
nabts.h — Bug 3
Added NABTS_FSS_MAX_PACKETS (68) and NABTS_FSS_MAX_NAPLPS (1896) constants, derived directly from §8.4.2.5.
hamming.c — Incidental fix
Fixed a pre-existing -Wsign-compare warning in str_parity(): int n changed to size_t n to match the size_t size parameter it's compared against. No behaviour change.
test_fixes.c — New file
23 regression tests covering all three fixes. Confirmed: 23/23 new tests pass, 126/126 original verify_spec tests still pass.

+++ hamming.c
@@ -51,8 +51,8 @@ void str_parity(uint8_t *str, size_t size, const char *format, ...)
     vsnprintf((char *)tmp, size+1, format, argList);
     va_end(argList);
 
-    int n;
-    for ( n=0; n<size; n++) {
+    size_t n;
+    for (n = 0; n < size; n++) {
         str[n] = parity(tmp[n]);
     }
 }
